Working on async part.

+++ src/state.rs
@@ -7,11 +7,11 @@ pub struct State<'a> {
     pub user: Cow<'a, str>,
     pub password: Cow<'a, str>,
     pub greeting_packet: GreetingPacket<'a>,
-    pub request_id: u32,
+    pub request_id: u64,
 }
 
 impl<'a> State<'a> {
-    pub fn get_id(&mut self) -> u32 {
+    pub fn get_id(&mut self) -> u64 {
         self.request_id += 1;
         self.request_id
     }
